% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citbi_raw.r
\docType{data}
\name{citbi_raw}
\alias{citbi_raw}
\title{Raw Citbi data}
\format{
An object of class \code{data.frame} with 30379 rows and 26 columns.
}
\usage{
citbi_raw
}
\description{
Original and uncleaned version of the Citbi Dataset
}
\keyword{datasets}
