% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_NA.r
\name{insert_NA}
\alias{insert_NA}
\title{Remove NA}
\usage{
insert_NA(x)
}
\arguments{
\item{x}{Is the citbi dataset}
}
\value{
Returns the citbi dataset with the NA's or values equivalent to NA removed or altered
}
\description{
Removes values that are counted as NA or actually NA from the non-renamed citbi dataset
}
