% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.Date.R
\name{is.Date}
\alias{is.Date}
\title{Check for date type}
\usage{
is.Date(x)
}
\arguments{
\item{x}{input to check}
}
\description{
This function checks if the input is a valid date type. It supports all date types, including \code{Date}, \code{POSIXct} and \code{POSIXlt}.
}
