% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_citbi.r
\name{rename_citibi}
\alias{rename_citibi}
\title{Rename columns}
\usage{
rename_citibi(x)
}
\arguments{
\item{x}{The citbi dataset}
}
\value{
The citbi dataset with certain columns renamed
}
\description{
Rename the columns of the citbi dataset and should occur last in the cleaning process
}
